Revert "Revert "initial scaffolding for requestBodyBuffer (#4755)" (#4771)"
This reverts commit 0290543f180487d6da8d35a44f1921dfb6748ac2.
